refactor: move mutations files into queries folder
