+++ src/features/offer/components/OfferFooter/OfferFooter.tsx
@@ -7,9 +7,9 @@ import { CineContentCTA } from 'features/offer/components/OfferCine/CineContentC
 import { useOfferCTA } from 'features/offer/components/OfferContent/OfferCTAProvider'
 import { StickyFooterContent } from 'features/offer/components/OfferContent/StickyFooterContent/StickyFooterContent'
 import { getIsAComingSoonOffer } from 'features/offer/helpers/getIsAComingSoonOffer'
-import { useAddReminderMutation } from 'features/offer/mutations/useAddReminderMutation'
-import { useDeleteReminderMutation } from 'features/offer/mutations/useDeleteReminderMutation'
 import { selectReminderByOfferId } from 'features/offer/queries/selectors/selectReminderByOfferId'
+import { useAddReminderMutation } from 'features/offer/queries/useAddReminderMutation'
+import { useDeleteReminderMutation } from 'features/offer/queries/useDeleteReminderMutation'
 import { useGetRemindersQuery } from 'features/offer/queries/useGetRemindersQuery'
 import { FavoriteProps } from 'features/offer/types'
 import { useRemoteConfigQuery } from 'libs/firebase/remoteConfig/queries/useRemoteConfigQuery'
