+++ src/features/offer/queries/types.ts
